Adjusted pantry inventory value when added to day and added alerts

+++ DailyPantry-cs329E/DailyPantry-cs329E/Model/StoredIngredient+CoreDataProperties.swift
@@ -18,7 +18,7 @@ extension StoredIngredient {
 
     @NSManaged public var name: String?
     @NSManaged public var catagory: String?
-    @NSManaged public var amountAvailable: Int16
+    @NSManaged public var amountAvailable: Double
     @NSManaged public var recipe: StoredRecipe?
 
 }
